Fix warning for type comparison

+++ src/serialize.h
@@ -561,7 +561,7 @@ public:
             if (idx >= (int32_t)size) {
                 throw std::ios_base::failure("out of bounds index");
             }
-            if (last != -1 && idx <= last) {
+            if (last != (size_t)-1 && idx <= (int32_t)last) {
                 throw std::ios_base::failure("offset overflow");
             }
             vec[idx] = true;
